top1k-rating-movie/requirements.txt
psycopg2-binary
pandas
sqlalchemy
numpy